(fix bug) complete integration test with clean up
