+++ server/spec/cleanUpDb.js
@@ -1,0 +1,19 @@
+var mysql = require('mysql');
+
+var connection = mysql.createConnection({
+  host     : 'localhost',
+  user     : 'root',
+  password : '',
+  database : 'carpool'
+});
+
+connection.connect();
+
+connection.query('DELETE Users.*, Trips.*, TripUsers.* FROM Users, Trips, TripUsers WHERE (Trips.eventfulId = "TestParty" OR Trips.eventfulId = "SecondTestParty") AND TripUsers.TripId = Trips.id AND Users.id = TripUsers.UserId;', function (error, result) {
+  if (error) {
+    throw error;
+  }
+  console.log('deleted ' + result.affectedRows + ' rows. All clean!');
+});
+
+connection.end();
